+++ packages/cf-deploy-config-inquirer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/cf-deploy-config-inquirer
 
+## 0.1.1
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.5.5
+
 ## 0.1.0
 
 ### Minor Changes