+++ packages/project-access/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/project-access
 
+## 1.28.8
+
+### Patch Changes
+
+-   Updated dependencies [73475e5]
+    -   @sap-ux/ui5-config@0.26.0
+
 ## 1.28.7
 
 ### Patch Changes